Add docs to exported types
